Booknstay/Server/book_n_stay/requirements.txt
asgiref==3.7.2
coverage==7.4.0
Django==5.0.1
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
mysqlclient==2.2.1
pillow==10.2.0
PyJWT==2.8.0
python-decouple==3.8
python-dotenv==1.0.0
typing_extensions==4.9.0
sqlparse==0.4.4
pytz==2023.3.post1
sqlparse==0.4.4
stripe==8.1.0
typing_extensions==4.9.0
gunicorn==21.2.0
xhtml2pdf==0.2.15